Adds fixes for failing tests. Adds safelink package to pager
