+++ packages/ndla-pager/package.json
@@ -20,7 +20,8 @@
     "es"
   ],
   "dependencies": {
-    "@ndla/core": "^0.6.16"
+    "@ndla/core": "^0.6.16",
+    "@ndla/safelink": "^0.0.2"
   },
   "peerDependencies": {
     "@emotion/core": "^10.0.9",
