+++ packages/ndla-ui/src/NDLAFilm/__tests__/isLetter-test.jsx
@@ -8,7 +8,7 @@
 
 /* eslint-env jest */
 
-import { isLetter } from '../AllMoviesAlphabetically';
+import { isLetter } from '../isLetter';
 
 test('Check is letter', () => {
   expect(isLetter('A')).toBe(true);
